+++ test/unit/model/node_test.dart
@@ -3,8 +3,6 @@ import 'package:fast_i18n/src/model/node.dart';
 import 'package:test/test.dart';
 
 void main() {
-  const localeEnum = 'AppLocale.en';
-
   group(ListNode, () {
     final interpolation = StringInterpolation.braces;
 
@@ -14,13 +12,11 @@ void main() {
           path: '',
           raw: 'Hello',
           interpolation: interpolation,
-          localeEnum: localeEnum,
         ),
         TextNode(
           path: '',
           raw: 'Hi',
           interpolation: interpolation,
-          localeEnum: localeEnum,
         ),
       ]);
 
@@ -33,13 +29,11 @@ void main() {
           path: '',
           raw: 'Hello',
           interpolation: interpolation,
-          localeEnum: localeEnum,
         ),
         TextNode(
           path: '',
           raw: 'Hi {name}',
           interpolation: interpolation,
-          localeEnum: localeEnum,
         ),
       ]);
 
@@ -53,13 +47,11 @@ void main() {
             path: '',
             raw: 'Hello',
             interpolation: interpolation,
-            localeEnum: localeEnum,
           ),
           TextNode(
             path: '',
             raw: 'Hi {name}',
             interpolation: interpolation,
-            localeEnum: localeEnum,
           ),
         ]),
         ListNode(path: '', entries: [
@@ -67,13 +59,11 @@ void main() {
             path: '',
             raw: 'Hello',
             interpolation: interpolation,
-            localeEnum: localeEnum,
           ),
           TextNode(
             path: '',
             raw: 'Hi {name}',
             interpolation: interpolation,
-            localeEnum: localeEnum,
           ),
         ]),
       ]);
@@ -89,13 +79,11 @@ void main() {
               path: '',
               raw: 'Hello',
               interpolation: interpolation,
-              localeEnum: localeEnum,
             ),
             TextNode(
               path: '',
               raw: 'Hi',
               interpolation: interpolation,
-              localeEnum: localeEnum,
             ),
           ]),
         ]),
@@ -105,13 +93,11 @@ void main() {
               path: '',
               raw: 'Hello',
               interpolation: interpolation,
-              localeEnum: localeEnum,
             ),
             TextNode(
               path: '',
               raw: 'Hi',
               interpolation: interpolation,
-              localeEnum: localeEnum,
             ),
           ]),
           ListNode(path: '', entries: [
@@ -119,13 +105,11 @@ void main() {
               path: '',
               raw: 'Hello',
               interpolation: interpolation,
-              localeEnum: localeEnum,
             ),
             TextNode(
               path: '',
               raw: 'Hi',
               interpolation: interpolation,
-              localeEnum: localeEnum,
             )
           ]),
         ]),
@@ -143,7 +127,6 @@ void main() {
               path: '',
               raw: 'Hi',
               interpolation: interpolation,
-              localeEnum: localeEnum,
             ),
           },
           isMap: false,
@@ -155,7 +138,6 @@ void main() {
               path: '',
               raw: 'Hi',
               interpolation: interpolation,
-              localeEnum: localeEnum,
             ),
           },
           isMap: false,
@@ -174,7 +156,6 @@ void main() {
               path: '',
               raw: 'Hi',
               interpolation: interpolation,
-              localeEnum: localeEnum,
             ),
           },
           isMap: true,
@@ -186,7 +167,6 @@ void main() {
               path: '',
               raw: 'Hi',
               interpolation: interpolation,
-              localeEnum: localeEnum,
             ),
           },
           isMap: true,
@@ -201,42 +181,42 @@ void main() {
     group(StringInterpolation.dart, () {
       test('no arguments', () {
         final test = 'No arguments';
-        final node = TextNode.test(test, StringInterpolation.dart, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.dart);
         expect(node.content, test);
         expect(node.params, <String>{});
       });
 
       test('one argument', () {
         final test = 'I have one argument named \$apple.';
-        final node = TextNode.test(test, StringInterpolation.dart, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.dart);
         expect(node.content, test);
         expect(node.params, {'apple'});
       });
 
       test('one duplicate argument', () {
         final test = 'This string has one \$argument and \$argument.';
-        final node = TextNode.test(test, StringInterpolation.dart, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.dart);
         expect(node.content, test);
         expect(node.params, {'argument'});
       });
 
       test('one argument at the beginning', () {
         final test = '\$test at the beginning.';
-        final node = TextNode.test(test, StringInterpolation.dart, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.dart);
         expect(node.content, test);
         expect(node.params, {'test'});
       });
 
       test('one escaped argument', () {
         final test = 'I have one argument named \\\$apple.';
-        final node = TextNode.test(test, StringInterpolation.dart, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.dart);
         expect(node.content, 'I have one argument named \\\$apple.'); // \$apple
         expect(node.params, <String>{});
       });
 
       test('one argument with link', () {
         final test = '\$apple is linked to @:wow!';
-        final node = TextNode.test(test, StringInterpolation.dart, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.dart);
         expect(node.content,
             '\$apple is linked to \${AppLocale.en.translations.wow}!');
         expect(node.params, {'apple'});
@@ -245,7 +225,7 @@ void main() {
       test('one argument and single dollars', () {
         final test =
             r'$ I have \$ one argument named $apple but also a $ and a $';
-        final node = TextNode.test(test, StringInterpolation.dart, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.dart);
         expect(node.content,
             r'\$ I have \$ one argument named $apple but also a \$ and a \$');
         expect(node.params, {'apple'});
@@ -253,8 +233,8 @@ void main() {
 
       test('with case', () {
         final test = r'Nice $cool_hi $wow ${yes} ${no_yes} @:hello_world.yes';
-        final node = TextNode.test(
-            test, StringInterpolation.dart, localeEnum, CaseStyle.camel);
+        final node =
+            TextNode.test(test, StringInterpolation.dart, CaseStyle.camel);
         expect(node.content,
             r'Nice $coolHi $wow ${yes} ${noYes} ${AppLocale.en.translations.hello_world.yes}');
         expect(node.params, {'coolHi', 'wow', 'yes', 'noYes'});
@@ -262,7 +242,7 @@ void main() {
 
       test('with links', () {
         final test = r'@:.c @:a @:hi @:wow. @:nice.cool';
-        final node = TextNode.test(test, StringInterpolation.dart, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.dart);
         expect(node.content,
             r'@:.c ${AppLocale.en.translations.a} ${AppLocale.en.translations.hi} ${AppLocale.en.translations.wow}. ${AppLocale.en.translations.nice.cool}');
         expect(node.params, <String>{});
@@ -272,40 +252,35 @@ void main() {
     group(StringInterpolation.braces, () {
       test('no arguments', () {
         final test = 'No arguments';
-        final node =
-            TextNode.test(test, StringInterpolation.braces, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.braces);
         expect(node.content, test);
         expect(node.params, <String>{});
       });
 
       test('one argument', () {
         final test = 'I have one argument named {apple}.';
-        final node =
-            TextNode.test(test, StringInterpolation.braces, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.braces);
         expect(node.content, 'I have one argument named \$apple.');
         expect(node.params, {'apple'});
       });
 
       test('one argument without space', () {
         final test = 'I have one argument named{apple}.';
-        final node =
-            TextNode.test(test, StringInterpolation.braces, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.braces);
         expect(node.content, 'I have one argument named\$apple.');
         expect(node.params, {'apple'});
       });
 
       test('one argument followed by underscore', () {
         final test = 'This string has one argument named {tom}_';
-        final node =
-            TextNode.test(test, StringInterpolation.braces, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.braces);
         expect(node.content, 'This string has one argument named \${tom}_');
         expect(node.params, {'tom'});
       });
 
       test('one argument followed by number', () {
         final test = 'This string has one argument named {tom}7';
-        final node =
-            TextNode.test(test, StringInterpolation.braces, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.braces);
         expect(node.content, 'This string has one argument named \${tom}7');
         expect(node.params, {'tom'});
       });
@@ -313,8 +288,7 @@ void main() {
       test('one argument with fake arguments', () {
         final test =
             r'$ I have one argument named {apple} but this is $fake. \$ $';
-        final node =
-            TextNode.test(test, StringInterpolation.braces, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.braces);
         expect(node.content,
             r'\$ I have one argument named $apple but this is \$fake. \$ \$');
         expect(node.params, {'apple'});
@@ -322,16 +296,14 @@ void main() {
 
       test('one escaped argument escaped', () {
         final test = 'I have one argument named \\{apple}.';
-        final node =
-            TextNode.test(test, StringInterpolation.braces, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.braces);
         expect(node.content, 'I have one argument named {apple}.');
         expect(node.params, <String>{});
       });
 
       test('one argument with link', () {
         final test = '{apple} is linked to @:wow!';
-        final node =
-            TextNode.test(test, StringInterpolation.braces, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.braces);
         expect(node.content,
             '\$apple is linked to \${AppLocale.en.translations.wow}!');
         expect(node.params, {'apple'});
@@ -339,8 +311,8 @@ void main() {
 
       test('with case', () {
         final test = r'Nice {cool_hi} {wow} {yes}a {no_yes}';
-        final node = TextNode.test(
-            test, StringInterpolation.braces, localeEnum, CaseStyle.camel);
+        final node =
+            TextNode.test(test, StringInterpolation.braces, CaseStyle.camel);
         expect(node.content, r'Nice $coolHi $wow ${yes}a $noYes');
         expect(node.params, {'coolHi', 'wow', 'yes', 'noYes'});
       });
@@ -349,56 +321,50 @@ void main() {
     group(StringInterpolation.doubleBraces, () {
       test('no arguments', () {
         final test = 'No arguments';
-        final node =
-            TextNode.test(test, StringInterpolation.doubleBraces, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.doubleBraces);
         expect(node.content, test);
         expect(node.params, <String>{});
       });
 
       test('one argument', () {
         final test = 'I have one argument named {{apple}}.';
-        final node =
-            TextNode.test(test, StringInterpolation.doubleBraces, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.doubleBraces);
         expect(node.content, 'I have one argument named \$apple.');
         expect(node.params, {'apple'});
       });
 
       test('one argument followed by underscore', () {
         final test = 'This string has one argument named {{tom}}_';
-        final node =
-            TextNode.test(test, StringInterpolation.doubleBraces, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.doubleBraces);
         expect(node.content, 'This string has one argument named \${tom}_');
         expect(node.params, {'tom'});
       });
 
       test('one argument followed by number', () {
         final test = 'This string has one argument named {{tom}}7';
-        final node =
-            TextNode.test(test, StringInterpolation.doubleBraces, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.doubleBraces);
         expect(node.content, 'This string has one argument named \${tom}7');
         expect(node.params, {'tom'});
       });
 
       test('one fake argument', () {
         final test = 'I have one argument named \\{apple}.';
-        final node =
-            TextNode.test(test, StringInterpolation.doubleBraces, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.doubleBraces);
         expect(node.content, test);
         expect(node.params, <String>{});
       });
 
       test('one escaped argument', () {
         final test = 'I have one argument named \\{{apple}}.';
-        final node =
-            TextNode.test(test, StringInterpolation.doubleBraces, localeEnum);
+        final node = TextNode.test(test, StringInterpolation.doubleBraces);
         expect(node.content, 'I have one argument named {{apple}}.');
         expect(node.params, <String>{});
       });
 
       test('with case', () {
         final test = r'Nice {{cool_hi}} {{wow}} {{yes}}a {{no_yes}}';
-        final node = TextNode.test(test, StringInterpolation.doubleBraces,
-            localeEnum, CaseStyle.camel);
+        final node = TextNode.test(
+            test, StringInterpolation.doubleBraces, CaseStyle.camel);
         expect(node.content, r'Nice $coolHi $wow ${yes}a $noYes');
         expect(node.params, {'coolHi', 'wow', 'yes', 'noYes'});
       });